Refactor for better package structure
