+++ src/main/java/com/monzo/queue/FrontierQueue.java
@@ -1,4 +1,4 @@
-package com.monzo;
+package com.monzo.queue;
 
 import java.util.concurrent.CompletableFuture;
