+++ src/test/java/com/monzo/config/CrawlerConfigTest.java
@@ -1,4 +1,4 @@
-package com.monzo;
+package com.monzo.config;
 
 import org.junit.jupiter.api.Test;
